+++ server/controllers/tutorials.js
@@ -12,7 +12,7 @@ export const createTutorial = async (req, res) => {
         res.status(201).json(newTutorial);
     } catch (error) {
         console.log(error)
-        res.status(409).json({ message: error.message });
+        res.status(409).json({ message: "Nie udało się utworzyć poradnika.", desc: error.message });
     }
 }
 
@@ -23,7 +23,7 @@ export const getTutorialsByUsername = async (req, res) => {
         const tutorials = await Tutorial.find({ username: username });
         res.status(200).json(tutorials.slice(0).reverse());
     } catch (error) {
-        console.log(error)
+        res.status(500).json({ message: "Błąd serwera. Spróbuj ponownie później.", desc: error.message });
     }
 }
 
@@ -31,11 +31,13 @@ export const deleteTutorialById = async (req, res) => {
     const { id } = req.params;
 
     try {
+        if(!mongoose.Types.ObjectId.isValid(id)) return res.status(404).json({ message: "Nie ma takiego posta z tym id" });
+
         await Tutorial.findByIdAndRemove(id);
 
         res.status(200).json({ message: "Usunięto poradnik pomyślnie" })
     } catch (error) {
-        console.log(error)
+        res.status(500).json({ message: "Błąd serwera. Spróbuj ponownie później.", desc: error.message });
     }
 }
 
@@ -43,11 +45,13 @@ export const getTutorialById = async (req, res) => {
     const { id } = req.params;
 
     try {
+        if(!mongoose.Types.ObjectId.isValid(id)) return res.status(404).json({ message: "Nie ma takiego posta z tym id" });
+
         const tutorial = await Tutorial.findById(id);
 
         res.status(200).json(tutorial)
     } catch (error) {
-        console.log(error)
+        res.status(500).json({ message: "Błąd serwera. Spróbuj ponownie później.", desc: error.message });
     }
 }
 
@@ -56,12 +60,14 @@ export const updateTutorial = async (req, res) => {
     const tutorial = req.body;
 
     try {
+        if(!mongoose.Types.ObjectId.isValid(id)) return res.status(404).json({ message: "Nie ma takiego posta z tym id" });
+
         await Tutorial.findByIdAndUpdate(id, tutorial, { new: true })
 
         const updatedTutorial = await Tutorial.findById(id)
 
         res.status(200).json(updatedTutorial)
     } catch (error) {
-        console.log(error)
+        res.status(500).json({ message: "Błąd serwera. Spróbuj ponownie później.", desc: error.message });
     }
 }
